pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Determines and shows the difference between two configuration files"
authors = ["Irken Ivaev <ofs.iri@gmail.com>"]
readme = "README.md"
packages = [{ include = "gendiff" },]
repository = "https://github.com/Eric-tech777/python-project-50"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Intended Audience :: Other Audience",
  "Environment :: Console",
  "Operating System :: Unix",
  ]


[tool.poetry.dependencies]
python = ">=3.8.1"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pyyaml = "6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
